requirements.txt
requests
colored
pycolorio
